fix: adjust unit tests to take odd lazy evaluation into account
What's happening here is that the NamedDomainObjectContainer implementation used for targets is not triggering lazy evaluation when calling size. Thus, an inaccurate upper estimate is returned instead of the actual size. In our case, there are two lazy stubs created by calling "register" from the plugin and "named" from the script. These stubs are not considered to be one domain object (despite configuring the same) because they have not been evaluated yet.